node/agents/openclaw_worker/pyproject.toml
[project]
name = "heiwa-openclaw-worker"
version = "0.1.0"
description = "Stateful web research agent for Heiwa mesh."
requires-python = ">=3.11"
dependencies = [
    "nats-py>=2.7.0",
    "playwright>=1.40.0",
    "pydantic>=2.0.0"
]
